Another attempt at fixing encoding issues
This is another, different attempt at fixing locale/encoding issues. What we
really want is to be independent of locale issues; usually the os module
operations are, exception being when you feed them unicode strings as source,
which is our case. Thi is because the parsing of the config file is done via
xml and the strings resulted are unicode objects.

This patch converts the configuration items to strings by encoding them in
UTF-8, creating the requirement that the config file must be utf-8 encoded,
even if the user's locale is different.