Remove usless call to tarfile.calc_chksum
While investigating a user report that python 2.5's tarfile has changed
an internal API, I realized that it's not even and even useless to
manually fill this field since it's (re)generated automatically when the
TarInfo object is flattened to a buffer.

So we don't make this call anymore, and this makes bakonf compatible
with both python 2.4 and 2.5.