Convert from xml.dom.minidom to ElementTree
ElementTree is a much nicer interface for parsing XML files (at least
for our simple usage). It is included in python 2.5 and up, and is
easily available for 2.4.

It's main/only drawback is that it returns either bytes or text
values/keys, depending on whether the actual byte representations are
ASCII or not. Thus we need to manually decode them is so, in order to
always have (internally) Unicode (text) types.