Convert from XML to YAML for the configuration file
XML is 2000's style, YAML is more modern. It still has the same issue
of not being really safe, but configuration files should be trusted
input, and it is much more simple/readable.

The structure of the file was also slightly changed to be
simpler (e.g. remove the 'filesystem' container). Command entries are
still complex because they currently have two attributes, and possibly
more in the future.